fix: widen previousMvir/Vvir/Vmax from float to double in core_build_model
These temporaries captured GALAXY struct float fields and were immediately
assigned to other float fields (infallMvir/Vvir/Vmax).  Using double avoids
a narrowing on read that could hide subtle precision differences if the
underlying fields are ever promoted.  No change to stored values since the
source and destination are still float.

+++ src/core_build_model.c
@@ -222,9 +222,9 @@ static int join_galaxies_of_progenitors(const int halonr, const int ngalstart, i
                 }
 
                 // remember properties from the last snapshot
-                const float previousMvir = galaxies[ngal].Mvir;
-                const float previousVvir = galaxies[ngal].Vvir;
-                const float previousVmax = galaxies[ngal].Vmax;
+                const double previousMvir = galaxies[ngal].Mvir;
+                const double previousVvir = galaxies[ngal].Vvir;
+                const double previousVmax = galaxies[ngal].Vmax;
 
                 if(prog == first_occupied) {
                     // update properties of this galaxy with physical properties of halo
